Add files via upload
The UCR Programme of Activities (PoA), “Gelson Solar Power Generation PoA, Gujarat ” (hereafter referred to as PoA) involves the installation of 520 small scale grid connected solar power plants with a total capacity of 3.886 MW in State: Gujarat, with the objective to contribute and promote generation of renewable energy in India. The commissioning date of the PoA is 29/11/2018.